Honor Retry-After in from_http
Teach from_http to reuse the Retry-After parsing and backoff logic
used by the shared HTTP client helpers.

This preserves retry diagnostics while making 429 handling follow the
server-provided delay and adds regression coverage for that path.
