+++ test/tests/operators/from_http/retry_status_429_retry_after.tql
@@ -1,0 +1,13 @@
+---
+fixtures:
+  - http_server:
+      expected_request:
+        count: 2
+        method: GET
+        path: /status/retry-429-after
+timeout: 60
+---
+
+from_http env("HTTP_FIXTURE_RETRY_429_AFTER_URL"), max_retry_count=1, retry_delay=1ms {
+  read_json
+}
